fix: WriteAngle wrote every negative angle as zero (B113)
Found by the first synthetic tests written under docs/MEASUREMENT-PLAN.md,
and it is exactly the shape that plan predicted.

svc_FixAngle's encoder was:

    writer.Write((uint)MathF.Round(degrees * (65536f / 360f)) & 0xFFFF, 16);

(uint) applied to a negative float does not wrap -- .NET's floating-point to
unsigned conversions SATURATE, so the value was clamped to zero before the
mask that looks like it handles the wrap ever saw it. A pitch of -30 degrees
(a player looking up, entirely ordinary) encoded as 0 and reproduced a player
looking dead level.

Valve's own bf_write::WriteBitAngle (tier1/bitbuf.cpp:551) gets it right by
going through a SIGNED int, where the mask does the two's-complement wrap:

    d = (int)( (fAngle / 360.0) * shift );
    d &= mask;

The fix is one cast, (uint)(int), which reproduces that: -90 encodes as 49152
and decodes as 270 -- the same direction by a different representative.

Why no existing test could have caught it, which is the interesting half:
CorpusMessageRoundTripTests reports 100% of payload bits reproduced, and that
was and remains true. ReadAngle is raw * 360f/65536f, so a demo-sourced angle
is always 0..360 and never re-enters the writer negative. The defect is
unreachable from any recording by construction, and reachable only from a
caller that builds a message in code -- a synthetic test, or the text-to-demo
compiler that is Phase 1's last item, which would have written flattened view
angles into every demo it produced.

Also adds NetMessageWriterTests: 12 synthetic round trips (object -> write ->
read -> object, so nothing is hand-encoded) over the simple messages, at
adversarial values rather than typical ones -- zero, field maxima, non-ASCII
strings, and the negative angles that found B113. Two things it documents that
cost a wrong prediction each: trailing bit padding decodes as net_NOP, so a
17-bit message gains a phantom Empty the corpus bit-comparison never sees; and
angles have no sign on the wire.

Measured on the box while writing this: cli mutation is 2m03s, 119 mutants,
76.30%, 16 survivors. Slot requested in MEASUREMENT-BOX-LOG.md.

Verified by sabotage: restoring the single (uint) cast reddens both
negative-angle tests and nothing else.

Gate green: core 1085, all floors held.

+++ managed/Tf2DemoSalvage.Core/Net/NetMessageWriter.cs
@@ -422,7 +422,16 @@ public static class NetMessageWriter
     /// loses the round trip on about half of all values: the decoded float is not exactly
     /// representable, so dividing it back lands just under the integer it came from and the cast
     /// drops a whole step.
+    ///
+    /// **The cast goes through a SIGNED int, and that is load-bearing rather than incidental.**
+    /// <c>bf_write::WriteBitAngle</c> (<c>tier1/bitbuf.cpp:551</c>) is
+    /// <c>int d = (int)((fAngle / 360.0) * shift); d &amp;= mask;</c> — signed, then masked, so a
+    /// negative angle wraps into its positive representative (-90 becomes 270, the same
+    /// direction). Casting the float straight to <c>uint</c> looks equivalent and is not: .NET's
+    /// floating-point to unsigned conversions SATURATE, so every negative angle became exactly
+    /// zero. A player looking up is a negative pitch, so this silently flattened the view angle it
+    /// was asked to reproduce. See RISKS B113.
     /// </remarks>
     private static void WriteAngle(BitWriter writer, float degrees) =>
-        writer.Write((uint)MathF.Round(degrees * (65536f / 360f)) & 0xFFFF, 16);
+        writer.Write((uint)(int)MathF.Round(degrees * (65536f / 360f)) & 0xFFFF, 16);
 }
